oauth_server/requirements.txt
Flask==3.0.2
google-auth-oauthlib==1.2.0
google-auth==2.27.0
PyJWT==2.8.0 